modificacion sumando
se modifico de parseFloat a parseInt segundo sumando

+++ suma.js
@@ -1,5 +1,5 @@
 const sum = (a, b) => {
-  return parseFloat(a) + parseFloat(b);
+  return parseFloat(a) + parseInt(b);
 };
 
 module.exports = sum;
